Added Dark Theme To Mission Control (#50)
* Renamed Navbar

* A usual commit

Nothing too see here, move along...

* Added Dark Theme To Mission Control

Problem is, you can't untoggle it. There are a few bugs that will be
fixed in the future.

* Finished Dark Theme For Mission Control

Now, there are cookies to set the darktheme being active or inactive.
The only way to get this to render is to entirely refresh the page.